[ etc ] move useRecentlyBookmarked and useCardsByGender
