+++ src/@components/BestPiicklePage/hooks/useCardsByGender.ts
@@ -1,9 +1,9 @@
 import useSWR from "swr";
 
-import { realReq } from "../../../../core/api/common/axios";
-import { PATH } from "../../../../core/api/common/constants";
-import { CardList } from "../../../../types/cardCollection";
-import { PiickleSWRResponse } from "../../../../types/remote/swr";
+import { realReq } from "../../../core/api/common/axios";
+import { PATH } from "../../../core/api/common/constants";
+import { CardList } from "../../../types/cardCollection";
+import { PiickleSWRResponse } from "../../../types/remote/swr";
 
 export function useCardsByGender(gender: "남" | "여") {
   const { data } = useSWR<PiickleSWRResponse<CardList[]>>(
